Add support for typing of SPARQL queries
Separate MyGlobal and Error definitions to separate files.
Add query expressions in QueryTypes and a framework for
typing SPARQL queries with the QueryTyper class, which is
used by the CheckerAnalyzer Component.
Updates Collector and Typedef Components to work with SPARQL
query typing.
Add the Evaluator class, which uses a constraint based
approach to calculate types for SPARQL queries.
Add SPARQLParser (which is based on the openrdf SPARQL parser)
to parse SPARQL queries and build query expressions.
Add tests for these components.